🔨 Add | 마이페이지 내 UI 그래프 구현 - 풀이 개수 그래프 구현 (2)

+++ Assets/Scripts/MyPage/UI/BarUIController.cs
@@ -18,7 +18,7 @@ public class BarUIController : MonoBehaviour
         this.text.text = date;
     }
 
-    public void SetBarSize(int normalizedRate)
+    public void SetBarSize(float normalizedRate)
     {
         this.barImage.rectTransform.sizeDelta = new Vector2(this.fixedWidth,this.fixedHight * normalizedRate);
     }
